server/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ML & NLP
torch==2.2.0
transformers==4.37.2
openai-whisper==20240314
sentence-transformers==2.3.1
pandas==2.1.4
numpy==1.26.4
scipy==1.12.0

# Speech & Audio
pyannote.audio==3.0.1
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10

# API Clients
groq==0.4.2
httpx==0.27.0
aiohttp==3.9.2

# Storage & File Handling
aiofiles==23.2.1
boto3==1.34.19  # AWS S3
python-multipart==0.0.6

# Configuration & Logging
python-dotenv==1.0.0
pyyaml==6.0.1
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-mock==0.3.0
faker==22.6.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
ruff==0.2.1

# Type Stubs (for mypy)
types-PyYAML==6.0.12.12
types-aiofiles==23.2.0.20240115
sqlalchemy-stubs==0.4

# Development
ipython==8.21.0
jupyter==1.0.0
debugpy==1.8.1

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
